UI re-arrange
